chore(aws_cloudwatch_metrics sink): Reorganize sink into a directory with test files
